+++ Moco/Moco/MocoTrajectory.cpp
@@ -29,6 +29,28 @@ using namespace OpenSim;
 const std::vector<std::string> MocoTrajectory::m_allowedKeys =
         {"states", "controls", "multipliers", "derivatives"};
 
+MocoTrajectory::MocoTrajectory(
+        std::vector<std::string> state_names,
+        std::vector<std::string> control_names,
+        std::vector<std::string> multiplier_names,
+        std::vector<std::string> parameter_names)
+        : m_state_names(std::move(state_names)),
+          m_control_names(std::move(control_names)),
+          m_multiplier_names(std::move(multiplier_names)),
+          m_parameter_names(std::move(parameter_names)) {}
+
+MocoTrajectory::MocoTrajectory(
+        std::vector<std::string> state_names,
+        std::vector<std::string> control_names,
+        std::vector<std::string> multiplier_names,
+        std::vector<std::string> derivative_names,
+        std::vector<std::string> parameter_names)
+        : m_state_names(std::move(state_names)),
+          m_control_names(std::move(control_names)),
+          m_multiplier_names(std::move(multiplier_names)),
+          m_derivative_names(std::move(derivative_names)),
+          m_parameter_names(std::move(parameter_names)) {}
+
 MocoTrajectory::MocoTrajectory(const SimTK::Vector& time,
         std::vector<std::string> state_names,
         std::vector<std::string> control_names,
@@ -953,67 +975,98 @@ void MocoTrajectory::randomize(bool add, const SimTK::Random& randGen) {
             SimTK::RowVector(0)); // TODO (parameters)
 }
 
-bool MocoTrajectory::isCompatible(
-        const MocoProblemRep& mp, bool throwOnError) const {
+bool MocoTrajectory::isCompatible(const MocoProblemRep& mp,
+        bool requireAccelerations, bool throwOnError) const {
     ensureUnsealed();
     // Slack variables might be solver dependent, so we can't include them in
     // the compatibility check.
 
-    auto mpsn = mp.createStateInfoNames();
-    std::sort(mpsn.begin(), mpsn.end());
-    auto mpcn = mp.createControlInfoNames();
-    std::sort(mpcn.begin(), mpcn.end());
-    auto mpmn = mp.createMultiplierInfoNames();
-    std::sort(mpmn.begin(), mpmn.end());
-    auto mppn = mp.createParameterNames();
-    std::sort(mppn.begin(), mppn.end());
+    const int debugLevel = Object::getDebugLevel();
 
-    auto sn(m_state_names);
-    std::sort(sn.begin(), sn.end());
+    auto compare = [&throwOnError, &debugLevel](
+            std::string varType, std::vector<std::string> trajNames,
+            std::vector<std::string> probNames,
+            std::string message = "") {
+        std::sort(trajNames.begin(), trajNames.end());
+        std::sort(probNames.begin(), probNames.end());
+        if (trajNames == probNames) return true;
+        if (!throwOnError && debugLevel <= 0) return false;
 
-    auto cn(m_control_names);
-    std::sort(cn.begin(), cn.end());
+        int sum = (int)trajNames.size() + (int)probNames.size();
 
-    auto mn(m_multiplier_names);
-    std::sort(mn.begin(), mn.end());
+        std::stringstream ss;
 
-    auto dn(m_derivative_names);
-    std::sort(dn.begin(), dn.end());
-
-    auto pn(m_parameter_names);
-    std::sort(pn.begin(), pn.end());
-
-    // Create the expected names for the derivative variable names.
-    std::vector<std::string> mpdn;
-    for (auto name : mpsn) {
-        auto leafpos = name.find("value");
-        if (leafpos != std::string::npos) {
-            name.replace(leafpos, name.size(), "accel");
-            mpdn.push_back(name);
+        // http://www.cplusplus.com/reference/algorithm/set_difference/
+        {
+            std::vector<std::string> inTrajNotProb(sum);
+            auto inTrajNotProbEnd = std::set_difference(trajNames.begin(),
+                    trajNames.end(), probNames.begin(), probNames.end(),
+                    inTrajNotProb.begin());
+            inTrajNotProb.resize(inTrajNotProbEnd - inTrajNotProb.begin());
+            ss << "The trajectory and provided problem are not compatible. ";
+            if (!inTrajNotProb.empty()) {
+                ss << "The following " << varType
+                   << " are in the trajectory but not the problem:\n";
+                for (const auto& name : inTrajNotProb) {
+                    ss << "  " << name << "\n";
+                }
+            }
         }
-    }
-    std::vector<std::string> mpcdn; // Component derivative names only.
+
+        {
+            std::vector<std::string> inProbNotTraj(sum);
+            auto inProbNotTrajEnd = std::set_difference(probNames.begin(),
+                    probNames.end(), trajNames.begin(), trajNames.end(),
+                    inProbNotTraj.begin());
+            inProbNotTraj.resize(inProbNotTrajEnd - inProbNotTraj.begin());
+            if (!inProbNotTraj.empty()) {
+                ss << "The following " << varType
+                   << " are in the problem but not the trajectory:\n";
+                for (const auto& name : inProbNotTraj) {
+                    ss << "  " << name << "\n";
+                }
+            }
+        }
+        ss << message << "\n";
+
+        if (debugLevel <= 0) {
+            throw Exception(__FILE__, __LINE__,
+                    "MocoTrajectory::isCompatible()", ss.str());
+        }
+        std::cout << ss.str() << std::flush;
+        return false;
+    };
+
+    auto mpsn = mp.createStateInfoNames();
+
+    if (!compare("state(s)", m_state_names, mpsn)) return false;
+    if (!compare("control(s)", m_control_names, mp.createControlInfoNames()))
+        return false;
+    if (!compare("multiplier(s)", m_multiplier_names,
+                mp.createMultiplierInfoNames()))
+        return false;
+    if (!compare("parameter(s)", m_parameter_names, mp.createParameterNames()))
+        return false;
+
+    std::vector<std::string> mpdn; // Component derivative names only.
     const auto& implicitComponentRefs = mp.getImplicitComponentReferencePtrs();
     for (const auto& compRef : implicitComponentRefs) {
         const auto& derivName =
                 compRef.second->getAbsolutePathString() + "/" + compRef.first;
         mpdn.push_back(derivName);
-        mpcdn.push_back(derivName);
     }
-    std::sort(mpdn.begin(), mpdn.end());
-    std::sort(mpcdn.begin(), mpcdn.end());
+    if (requireAccelerations) {
+        // Create the expected names for the derivative variables.
+        for (auto name : mpsn) {
+            auto leafpos = name.find("value");
+            if (leafpos != std::string::npos) {
+                name.replace(leafpos, name.size(), "accel");
+                mpdn.push_back(name);
+            }
+        }
+    }
 
-    bool compatible = mpsn == sn && mpcn == cn && mpmn == mn &&
-                      // It's okay to not have any multibody dynamics 
-                      // derivatives (for solving the problem with an explicit 
-                      // dynamics mode).
-                      (mpcdn == dn || mpdn == dn) && mppn == pn;
-
-    // TODO more detailed error message specifying exactly what's different.
-    OPENSIM_THROW_IF(!compatible && throwOnError, Exception,
-            "Trajectory and provided problem are not compatible.");
-
-    return compatible;
+    return compare("derivative(s)", m_derivative_names, mpdn);
 }
 
 bool MocoTrajectory::isNumericallyEqual(