+++ Moco/Moco/MocoTrajectory.h
@@ -98,6 +98,20 @@ accelerations. These are stored in the trajectory as derivative variables. */
 class OSIMMOCO_API MocoTrajectory {
 public:
     MocoTrajectory() = default;
+    /// Create a trajectory with no data. To add data, use setNumTimes(),
+    /// setTime(), and the other setters.
+    MocoTrajectory(std::vector<std::string> state_names,
+            std::vector<std::string> control_names,
+            std::vector<std::string> multiplier_names,
+            std::vector<std::string> parameter_names);
+    /// Create a trajectory (including columns for derivatives) with no data.
+    /// To add data, use setNumTimes(), setTime(), and the other setters.
+    MocoTrajectory(
+            std::vector<std::string> state_names,
+            std::vector<std::string> control_names,
+            std::vector<std::string> multiplier_names,
+            std::vector<std::string> derivative_names,
+            std::vector<std::string> parameter_names);
     MocoTrajectory(const SimTK::Vector& time,
             std::vector<std::string> state_names,
             std::vector<std::string> control_names,
@@ -490,11 +504,18 @@ public:
     /// @name Comparisons
     /// @{
 
-    /// Do the state and control names in this trajectory match those in the
-    /// problem? This may not catch all possible incompatibilities.
-    /// The problem and this trajectory can still be compatible even if the
-    /// trajectory contains no derivative columns.
-    bool isCompatible(const MocoProblemRep&, bool throwOnError = false) const;
+    /// Do the state, control, multiplier, derivative, and parameter names in
+    /// this trajectory match those in the problem? This may not catch all
+    /// possible incompatibilities. If the trajectory should have generalized
+    /// accelerations (for implicit multibody dynamics mode), set
+    /// requireAccelerations to true.
+    /// To throw an exception with a detailed message if the problem is not
+    /// compatible, pass throwOnError as true. To get the detailed message
+    /// without an exception, set the Object debug level to a level greater than
+    /// 0 (e.g., `Object::setDebugLevel(1)` in C++, and
+    /// `OpenSimObject.setDebugLevel(1)` in MATLAB/Python).
+    bool isCompatible(const MocoProblemRep&, bool requireAccelerations = false,
+            bool throwOnError = false) const;
     /// Check if this trajectory is numerically equal to another trajectory.
     /// This uses SimTK::Test::numericallyEqual() internally.
     /// Accordingly, the tolerance is both a relative and absolute tolerance