+++ Moco/Tests/testMocoInterface.cpp
@@ -1218,8 +1218,7 @@ TEMPLATE_TEST_CASE("Guess", "", MocoTropterSolver, MocoCasADiSolver) {
     {
         MocoTrajectory explicitGuess = ms.createGuess();
         ms.set_multibody_dynamics_mode("implicit");
-        ms.setGuess(explicitGuess);
-        CHECK_THROWS_WITH(study.solve(),
+        CHECK_THROWS_WITH(ms.setGuess(explicitGuess),
             Catch::Contains(
                 "'multibody_dynamics_mode' set to 'implicit' and coordinate states "
                 "exist in the guess, but no coordinate accelerations were "
@@ -1544,6 +1543,55 @@ TEST_CASE("MocoTrajectory") {
     testCompareParametersRMS(5, 0.5, {"p1", "p2"});
     // Compare a lot of parameters.
     testCompareParametersRMS(100, 0.5);
+}
+
+TEST_CASE("MocoTrajectory isCompatible") {
+    std::cout.rdbuf(LogManager::cout.rdbuf());
+    std::cout.rdbuf(LogManager::cout.rdbuf());
+    MocoProblem problem;
+    problem.setModel(createSlidingMassModel());
+    problem.setTimeBounds(MocoInitialBounds(0), MocoFinalBounds(0, 10));
+    problem.setStateInfo("/slider/position/value", MocoBounds(0, 1),
+            MocoInitialBounds(0), MocoFinalBounds(1));
+    problem.setStateInfo("/slider/position/speed", {-100, 100}, 0, 0);
+
+    MocoProblemRep rep = problem.createRep();
+
+    CHECK(MocoTrajectory({"/slider/position/value", "/slider/position/speed"},
+            {"/actuator"}, {}, {}).isCompatible(rep, false, true));
+    CHECK(!MocoTrajectory({"/slider/position/value", "/slider/position/speed"},
+            {"/actuator"}, {}, {}).isCompatible(rep, true));
+
+    CHECK(MocoTrajectory({"/slider/position/value", "/slider/position/speed"},
+            {"/actuator"}, {}, {"/slider/position/accel"}, {}).isCompatible(rep, true));
+
+    CHECK(!MocoTrajectory({}, {}, {}, {}).isCompatible(rep));
+    CHECK(!MocoTrajectory({"/slider/position/value", "/slider/position/speed",
+                          "nonexistent"},
+            {"/actuator"}, {}, {}).isCompatible(rep));
+    CHECK(!MocoTrajectory({"/slider/position/value", "/slider/position/speed"},
+            {"/actuator"}, {"nonexistent"}, {}).isCompatible(rep));
+
+    CHECK_THROWS_WITH(
+            MocoTrajectory({"/slider/position/value",
+                             "/slider/position/speed"},
+                              {"/actuator"}, {}, {})
+                               .isCompatible(rep, true, true),
+            Catch::Contains("accel"));
+    CHECK_THROWS_WITH(
+            MocoTrajectory({}, {}, {}, {}).isCompatible(rep, false, true),
+            Catch::Contains("position"));
+    CHECK_THROWS_WITH(
+            MocoTrajectory({"/slider/position/value", "/slider/position/speed",
+                             "nonexistent"},
+                    {"/actuator"}, {}, {})
+                     .isCompatible(rep, false, true),
+            Catch::Contains("nonexistent"));
+    CHECK_THROWS_WITH(
+            MocoTrajectory({"/slider/position/value", "/slider/position/speed"}, {"/actuator"},
+            {"nonexistent"}, {})
+                               .isCompatible(rep, false, true),
+            Catch::Contains("nonexistent"));
 }
 
 TEST_CASE("MocoTrajectory randomize") {